curddjango/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-rest==0.8.7
djangorestframework==3.13.1
psycopg2==2.9.3
#pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
